pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "modeljury-env"
version = "0.1.0"
description = "An OpenEnv environment where AI agents learn to evaluate other LLMs"
authors = [{name = "Ayush Raj Chourasia"}]
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
]

[project.scripts]
server = "server.app.main:main"
